requirements.txt
# Image Processing Toolkit Dependencies

# Core image processing libraries
Pillow>=10.0.0
piexif>=1.1.3
numpy>=1.21.0
opencv-python>=4.5.0

# Image quality metrics
scikit-image>=0.19.0

# Visualization and plotting
matplotlib>=3.5.0

# Data analysis and CSV export
pandas>=1.5.0

# Noise generation for synthetic images
noise>=1.2.2

# WebP and AVIF support
# WebP: Native Pillow support
# AVIF: ImageMagick with AVIF support (install separately)
pillow-heif>=0.22.0  # Optional: for reading AVIF files in validation

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0

# Development dependencies (optional)
black>=22.0.0
flake8>=5.0.0